pyproject.toml
[project]
name = "fastapi-celery-rabbitmq-redis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery[redis]>=5.6.2",
    "langchain-azure-ai>=1.0.0",
    "langchain-openai>=1.1.9",
    "python-dotenv>=1.2.1",
]
